watcher/requirements.txt
psutil==5.9.8
pygetwindow==0.0.9
keyboard==0.13.5
pystray==0.19.5
Pillow==10.3.0
python-dotenv==1.0.1
ctypes